Rename bbox_v2 to boxes
